requirements.txt
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0
pandas>=2.2.0
scikit-learn>=1.4.0
xgboost>=2.0.0
sentence-transformers>=2.5.0
numpy>=1.26.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
